Reformat files so group can add components
